requirements.txt
IMAPClient==3.0.1
python-dotenv==1.0.1
